Goole shearch some changes
