+++ src/components/ImageFilter/ImageFilterTool.css
@@ -1,0 +1,27 @@
+.image-filter-tools {
+    margin-top: 17px;
+}
+
+.image-filter-tools input {
+    cursor: pointer;
+    width: 100%;
+}
+.apply-button{
+    padding: 6px 30px;
+    background: #1462be;
+    font-family: sans-serif;
+    color: #fff;
+    border: 1px solid #1462be;
+    border-radius: .25rem;
+}
+.apply-button:hover{
+    background: #edf0f2;
+    color: #181b7dd6;
+    transition: 0.3s
+}
+
+input[type='color'] {
+    width: 80%;
+    margin: 0 auto;
+}
+
